WispAPI/requirements.txt
# Web Framework
fastapi[standard]
uvicorn[standard]==0.24.0

# Database & ORM
supabase==2.7.4

# HTTP Client for API integrations
httpx==0.25.2
aiohttp==3.9.1

# Environment & Configuration
python-dotenv==1.0.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Development & Testing
pytest==7.4.3
black==23.11.0
flake8==6.1.0

pydantic-settings

# Utilities
python-dateutil==2.8.2